requirements.txt
google-generativeai
python-dotenv
pillow
pyautogui 